GenMySSH 2.0!
To prevent any possibility of reverse-engineering I implemented the Knuth Shuffle algorithm on the array of key characters so that the mouse coordinates that eventually translated into index numbers of the key array would never pull the same pattern upon unique page visit. 

I also added a button that copies a random set of 16 characters within the key to generate a secure password.